requirements.txt
aenum==3.1.15
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
APScheduler==3.10.4
async-timeout==4.0.3
attrs==23.2.0
bidict==0.23.1
blinker==1.7.0
Brotli==1.1.0
CacheControl==0.14.0
cachelib==0.13.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
Deprecated==1.2.14
dnspython==2.6.1
filelock==3.13.4
firebase-admin==6.5.0
Flask==3.0.3
Flask-Compress==1.14
Flask-Login==0.6.3
Flask-ReCaptcha==0.4.2
Flask-Session==0.8.0
Flask-SocketIO==5.3.6
frozenlist==1.4.1
future==1.0.0
google-api-core==2.18.0
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
httplib2==0.22.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
line-bot-sdk==3.11.0
lml==0.1.0
lxml==5.2.1
MarkupSafe==2.1.5
msgpack==1.0.8
msgspec==0.18.6
multidict==6.0.5
oauthlib==3.2.2
orjson==3.10.1
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
pyexcel-ezodf==0.3.4
pyexcel-io==0.6.6
pyexcel-ods3==0.6.1
PyJWT==2.8.0
pymongo==4.6.3
PyMuPDF==1.24.1
PyMuPDFb==1.24.1
pyparsing==3.1.2
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-engineio==4.9.0
python-socketio==5.11.2
pytz==2024.1
qrcode==7.4.2
reportlab==4.1.0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
simple-websocket==1.0.0
six==1.16.0
typing_extensions==4.11.0
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==3.0.2
wrapt==1.16.0
wsproto==1.2.0
xds-protos==1.62.1
yarl==1.9.4
gevent
gevent-websocket
apscheduler